Add a confirmation failsafe before deleting files
The function `dir`, used to determine MKFILE_DIR can take multiple
parameters. As a result, if this repository is located in a path
that contains a white space, the result of `abspath` is interpreted
as multiple parameters, because Make uses white spaces as separator
or delimiter. Quoting does not help. Spaces as such only work in
shell commands within target bodies.

So, the chosen workaround is to add a failsafe to ask the user for
confirmation to give them a chance to double check what is actually
being deleted.

* added a warning at the top of the readme to raise awareness